refactor: consolidate test suite — remove 41 redundant tests
Critic-validated audit found significant redundancy across test files.
This commit removes duplicates while preserving 75.6% statement coverage.

Changes:
- Delete src/router.test.ts (21 tests) — 16 duplicated formatting.test.ts;
  migrate 6 unique tests (routeOutbound, findChannel, stripInternalTags)
- Trim setup/platform.test.ts from 13 to 5 real-environment smoke tests;
  add isRoot branching tests to platform-expanded.test.ts
- Move 3 computeNextRun tests to task-scheduler-expanded.test.ts;
  keep integration invalid-folder test in original (real DB + scheduler)
- Trim test/e2e/mocked.test.ts from 7 to 1 comprehensive pipeline test
  covering multi-channel routing and content escaping
- Trim test/chaos/resilience.test.ts from 15 to 6 unique boundary tests
- Remove 3 duplicate task CRUD tests from db-expanded.test.ts

Result: 430 tests (was 471), 75.59% coverage (unchanged), 0 TS errors
