+++ src/task-scheduler-expanded.test.ts
@@ -60,6 +60,7 @@ import { resolveGroupFolderPath } from './group-folder.js';
 import {
   startSchedulerLoop,
   _resetSchedulerLoopForTests,
+  computeNextRun,
   SchedulerDependencies,
 } from './task-scheduler.js';
 import { ScheduledTask } from './types.js';
@@ -310,6 +311,79 @@ describe('task-scheduler expanded', () => {
           error: 'spawn error',
         }),
       );
+    });
+  });
+
+  describe('computeNextRun', () => {
+    it('anchors interval tasks to scheduled time to prevent drift', () => {
+      const scheduledTime = new Date(Date.now() - 2000).toISOString();
+      const task = {
+        id: 'drift-test',
+        group_folder: 'test',
+        chat_jid: 'test@g.us',
+        prompt: 'test',
+        schedule_type: 'interval' as const,
+        schedule_value: '60000',
+        context_mode: 'isolated' as const,
+        next_run: scheduledTime,
+        last_run: null,
+        last_result: null,
+        status: 'active' as const,
+        created_at: '2026-01-01T00:00:00.000Z',
+      };
+
+      const nextRun = computeNextRun(task);
+      expect(nextRun).not.toBeNull();
+
+      const expected = new Date(scheduledTime).getTime() + 60000;
+      expect(new Date(nextRun!).getTime()).toBe(expected);
+    });
+
+    it('returns null for once-tasks', () => {
+      const task = {
+        id: 'once-test',
+        group_folder: 'test',
+        chat_jid: 'test@g.us',
+        prompt: 'test',
+        schedule_type: 'once' as const,
+        schedule_value: '2026-01-01T00:00:00.000Z',
+        context_mode: 'isolated' as const,
+        next_run: new Date(Date.now() - 1000).toISOString(),
+        last_run: null,
+        last_result: null,
+        status: 'active' as const,
+        created_at: '2026-01-01T00:00:00.000Z',
+      };
+
+      expect(computeNextRun(task)).toBeNull();
+    });
+
+    it('skips missed intervals without infinite loop', () => {
+      const ms = 60000;
+      const missedBy = ms * 10;
+      const scheduledTime = new Date(Date.now() - missedBy).toISOString();
+
+      const task = {
+        id: 'skip-test',
+        group_folder: 'test',
+        chat_jid: 'test@g.us',
+        prompt: 'test',
+        schedule_type: 'interval' as const,
+        schedule_value: String(ms),
+        context_mode: 'isolated' as const,
+        next_run: scheduledTime,
+        last_run: null,
+        last_result: null,
+        status: 'active' as const,
+        created_at: '2026-01-01T00:00:00.000Z',
+      };
+
+      const nextRun = computeNextRun(task);
+      expect(nextRun).not.toBeNull();
+      expect(new Date(nextRun!).getTime()).toBeGreaterThan(Date.now());
+      const offset =
+        (new Date(nextRun!).getTime() - new Date(scheduledTime).getTime()) % ms;
+      expect(offset).toBe(0);
     });
   });
 });
