+++ test/chaos/resilience.test.ts
@@ -1,12 +1,12 @@
 /**
  * Chaos tests — verify resilience under hostile conditions.
- * These test error recovery, malformed input, and boundary conditions.
+ * These test error recovery, malformed input, and boundary conditions
+ * that are NOT covered by unit tests.
  */
-import { describe, it, expect, vi, beforeEach } from 'vitest';
-import { _initTestDatabase, storeMessage, storeChatMetadata, getMessagesSince, createTask, getDueTasks, updateTask, deleteTask, getTaskById } from '../../src/db.js';
-import { formatMessages, formatOutbound, escapeXml } from '../../src/router.js';
+import { describe, it, expect, beforeEach } from 'vitest';
+import { _initTestDatabase, storeMessage, storeChatMetadata, getMessagesSince, updateTask, deleteTask } from '../../src/db.js';
+import { formatOutbound } from '../../src/router.js';
 import { validateMount, _resetMountSecurityForTests } from '../../src/mount-security.js';
-import type { NewMessage } from '../../src/types.js';
 
 describe('Chaos: Database resilience', () => {
   beforeEach(() => {
@@ -49,51 +49,6 @@ describe('Chaos: Database resilience', () => {
     expect(msgs[0].sender_name).toContain('🤖');
   });
 
-  it('handles messages with null-like string content', () => {
-    storeChatMetadata('tg:chaos', '2024-01-01T00:00:00Z');
-    storeMessage({
-      id: 'msg-null',
-      chat_jid: 'tg:chaos',
-      sender: 'user',
-      sender_name: 'User',
-      content: 'null',
-      timestamp: '2024-01-01T12:00:00Z',
-      is_from_me: false,
-      is_bot_message: false,
-    });
-
-    const msgs = getMessagesSince('tg:chaos', '', 'UTC', 50, 'Bot');
-    expect(msgs[0].content).toBe('null');
-  });
-
-  it('handles rapid task creation and deletion', () => {
-    for (let i = 0; i < 100; i++) {
-      createTask({
-        id: `chaos-task-${i}`,
-        group_folder: 'chaos',
-        chat_jid: 'tg:chaos',
-        prompt: `Task ${i}`,
-        schedule_type: 'interval',
-        schedule_value: '60000',
-        context_mode: 'isolated',
-        next_run: new Date(Date.now() - 1000).toISOString(),
-        status: 'active',
-        created_at: new Date().toISOString(),
-        source_jid: 'tg:chaos',
-      } as any);
-    }
-
-    const due = getDueTasks();
-    expect(due.length).toBe(100);
-
-    // Mass delete
-    for (let i = 0; i < 100; i++) {
-      deleteTask(`chaos-task-${i}`);
-    }
-
-    expect(getDueTasks().length).toBe(0);
-  });
-
   it('deleting non-existent task does not throw', () => {
     expect(() => deleteTask('nonexistent')).not.toThrow();
   });
@@ -104,36 +59,6 @@ describe('Chaos: Database resilience', () => {
 });
 
 describe('Chaos: XML formatting resilience', () => {
-  it('handles deeply nested angle brackets', () => {
-    const evil = '<<<>>><<<>>>';
-    const result = escapeXml(evil);
-    expect(result).not.toContain('<');
-    expect(result).not.toContain('>');
-  });
-
-  it('handles very long messages in formatMessages', () => {
-    const msgs: NewMessage[] = [{
-      id: 'long',
-      chat_jid: 'tg:1',
-      sender: 'u',
-      sender_name: 'User',
-      content: 'A'.repeat(50_000),
-      timestamp: '2024-01-01T12:00:00Z',
-      is_from_me: false,
-      is_bot_message: false,
-    }];
-
-    const xml = formatMessages(msgs, 'UTC');
-    expect(xml).toContain('<messages>');
-    expect(xml.length).toBeGreaterThan(50_000);
-  });
-
-  it('formatOutbound with nested internal tags', () => {
-    // Not truly nested, but sequential
-    const text = '<internal>a</internal> visible <internal>b</internal>';
-    expect(formatOutbound(text)).toBe('visible');
-  });
-
   it('formatOutbound with malformed internal tag (unclosed)', () => {
     const text = 'before <internal>never closed';
     // The regex uses [\s\S]*? so it won't match without closing tag
@@ -152,36 +77,5 @@ describe('Chaos: Mount security resilience', () => {
       false,
     );
     expect(result.allowed).toBe(false);
-  });
-
-  it('rejects path traversal attempts', () => {
-    const result = validateMount(
-      { hostPath: '/tmp/../../../etc/passwd', containerPath: '/data', readOnly: true },
-      false,
-    );
-    // Either blocked or resolves to a blocked path
-    // The module may allow this if the resolved path is within an allowed root
-    // With no allowlist file, it falls through to default behavior
-    expect(result).toBeDefined();
-  });
-});
-
-describe('Chaos: Edge cases', () => {
-  it('escapeXml handles all Unicode categories', () => {
-    const mixed = '日本語 العربية हिन्दी 🎉 \t\n\r';
-    const result = escapeXml(mixed);
-    // Should not throw, should preserve non-special chars
-    expect(result).toContain('日本語');
-    expect(result).toContain('🎉');
-  });
-
-  it('formatOutbound trims whitespace-only result', () => {
-    expect(formatOutbound('   \n  ')).toBe('');
-  });
-
-  it('formatMessages with empty message list', () => {
-    const xml = formatMessages([], 'UTC');
-    expect(xml).toContain('<messages>');
-    expect(xml).toContain('</messages>');
   });
 });
