+++ test/e2e/mocked.test.ts
@@ -6,16 +6,16 @@
  * parse output → deliver response via mock channel
  */
 import { describe, it, expect, vi, beforeEach } from 'vitest';
-import { formatMessages, formatOutbound, routeOutbound, findChannel } from '../../src/router.js';
-import { escapeXml } from '../../src/router.js';
+import { formatMessages, formatOutbound, routeOutbound } from '../../src/router.js';
 import { _initTestDatabase, storeMessage, storeChatMetadata, getMessagesSince } from '../../src/db.js';
-import type { Channel, NewMessage, RegisteredGroup } from '../../src/types.js';
+import type { Channel, NewMessage } from '../../src/types.js';
 
 // Mock channel that records sends
 function createMockChannel(jidPrefix: string): Channel & { sent: Array<{jid: string; text: string}> } {
   const sent: Array<{jid: string; text: string}> = [];
   return {
     sent,
+    name: `mock-${jidPrefix}`,
     connect: vi.fn(async () => {}),
     disconnect: vi.fn(async () => {}),
     ownsJid: (jid: string) => jid.startsWith(jidPrefix),
@@ -27,17 +27,22 @@ function createMockChannel(jidPrefix: string): Channel & { sent: Array<{jid: str
 }
 
 describe('E2E Mocked: Full message lifecycle', () => {
-  let channel: ReturnType<typeof createMockChannel>;
+  let tgChannel: ReturnType<typeof createMockChannel>;
+  let waChannel: ReturnType<typeof createMockChannel>;
 
   beforeEach(() => {
     _initTestDatabase();
-    channel = createMockChannel('tg:');
+    tgChannel = createMockChannel('tg:');
+    waChannel = createMockChannel('wa:');
   });
 
-  it('processes a message through the full pipeline', async () => {
-    // Step 1: Store a chat and a message (simulates channel ingestion)
-    storeChatMetadata('tg:group1', '2024-01-01T00:00:00Z', 'Test Group');
-    const msg: NewMessage = {
+  it('processes messages through the full pipeline with multi-channel routing and content escaping', async () => {
+    // Step 1: Set up two chats on different channels
+    storeChatMetadata('tg:group1', '2024-01-01T00:00:00Z', 'TG Group');
+    storeChatMetadata('wa:group2', '2024-01-01T00:00:00Z', 'WA Group');
+
+    // Step 2: Store messages including dangerous content
+    const msg1: NewMessage = {
       id: 'msg-001',
       chat_jid: 'tg:group1',
       sender: 'user-1',
@@ -47,127 +52,55 @@ describe('E2E Mocked: Full message lifecycle', () => {
       is_from_me: false,
       is_bot_message: false,
     };
-    storeMessage(msg);
-
-    // Step 2: Retrieve messages (simulates getNewMessages)
-    const messages = getMessagesSince('tg:group1', '', 'UTC', 50, 'Bot');
-    expect(messages.length).toBeGreaterThanOrEqual(1);
-    expect(messages[0].content).toContain('@Bot');
-
-    // Step 3: Format for container input
-    const xml = formatMessages(messages, 'UTC');
-    expect(xml).toContain('<messages>');
-    expect(xml).toContain('sender="Alice"');
-    expect(xml).toContain('@Bot help me');
-
-    // Step 4: Simulate container output
-    const containerOutput = {
-      status: 'success' as const,
-      result: 'Here is the help you need! <internal>reasoning here</internal>',
-      newSessionId: 'sess-123',
-    };
-
-    // Step 5: Format outbound (strip internal tags)
-    const cleaned = formatOutbound(containerOutput.result);
-    expect(cleaned).toBe('Here is the help you need!');
-    expect(cleaned).not.toContain('<internal>');
-
-    // Step 6: Route to channel
-    await routeOutbound([channel], 'tg:group1', cleaned);
-
-    // Step 7: Verify delivery
-    expect(channel.sent).toHaveLength(1);
-    expect(channel.sent[0].jid).toBe('tg:group1');
-    expect(channel.sent[0].text).toBe('Here is the help you need!');
-  });
-
-  it('blocks delivery when no channel owns the JID', async () => {
-    const waChannel = createMockChannel('wa:'); // wrong prefix
-    expect(() => routeOutbound([waChannel], 'tg:group1', 'test')).toThrow(
-      'No channel for JID',
-    );
-  });
-
-  it('handles empty agent response gracefully', () => {
-    const result = formatOutbound('<internal>only thinking</internal>');
-    expect(result).toBe('');
-    // If empty, we wouldn't route — verified by skipping routeOutbound
-  });
-
-  it('multi-turn: messages accumulate and are formatted together', () => {
-    storeChatMetadata('tg:group1', '2024-01-01T00:00:00Z');
-    storeMessage({
-      id: 'msg-1',
-      chat_jid: 'tg:group1',
-      sender: 'user-1',
-      sender_name: 'Alice',
-      content: 'First message',
-      timestamp: '2024-01-01T12:00:00Z',
-      is_from_me: false,
-      is_bot_message: false,
-    });
-    storeMessage({
-      id: 'msg-2',
+    const msg2: NewMessage = {
+      id: 'msg-002',
       chat_jid: 'tg:group1',
       sender: 'user-2',
-      sender_name: 'Bob',
-      content: 'Second message',
+      sender_name: '<script>alert(1)</script>',
+      content: 'a & b < c',
       timestamp: '2024-01-01T12:01:00Z',
       is_from_me: false,
       is_bot_message: false,
-    });
+    };
+    storeMessage(msg1);
+    storeMessage(msg2);
 
+    // Step 3: Retrieve and verify multiple messages
     const messages = getMessagesSince('tg:group1', '', 'UTC', 50, 'Bot');
+    expect(messages.length).toBeGreaterThanOrEqual(2);
+
+    // Step 4: Format for container — verify XML escaping
     const xml = formatMessages(messages, 'UTC');
-
-    expect(xml).toContain('Alice');
-    expect(xml).toContain('Bob');
-    expect(xml).toContain('First message');
-    expect(xml).toContain('Second message');
-  });
-
-  it('XML-escapes dangerous content in messages', () => {
-    storeChatMetadata('tg:group1', '2024-01-01T00:00:00Z');
-    storeMessage({
-      id: 'msg-xss',
-      chat_jid: 'tg:group1',
-      sender: 'attacker',
-      sender_name: '<script>alert(1)</script>',
-      content: 'a & b < c',
-      timestamp: '2024-01-01T12:00:00Z',
-      is_from_me: false,
-      is_bot_message: false,
-    });
-
-    const messages = getMessagesSince('tg:group1', '', 'UTC', 50, 'Bot');
-    const xml = formatMessages(messages, 'UTC');
-
+    expect(xml).toContain('<messages>');
+    expect(xml).toContain('sender="Alice"');
     expect(xml).not.toContain('<script>');
     expect(xml).toContain('&lt;script&gt;');
     expect(xml).toContain('a &amp; b &lt; c');
-  });
-});
 
-describe('E2E Mocked: Channel routing', () => {
-  it('routes to correct channel among multiple', async () => {
-    const tgChannel = createMockChannel('tg:');
-    const waChannel = createMockChannel('wa:');
+    // Step 5: Simulate container output with internal reasoning
+    const containerOutput = {
+      status: 'success' as const,
+      result: 'Here is help! <internal>reasoning here</internal>',
+      newSessionId: 'sess-123',
+    };
 
-    await routeOutbound([tgChannel, waChannel], 'tg:123', 'for telegram');
-    await routeOutbound([tgChannel, waChannel], 'wa:456', 'for whatsapp');
+    // Step 6: Format outbound — strip internal tags
+    const cleaned = formatOutbound(containerOutput.result);
+    expect(cleaned).toBe('Here is help!');
+    expect(cleaned).not.toContain('<internal>');
 
+    // Step 7: Route to correct channel among multiple
+    await routeOutbound([tgChannel, waChannel], 'tg:group1', cleaned);
+
+    // Step 8: Verify delivery to correct channel only
     expect(tgChannel.sent).toHaveLength(1);
+    expect(waChannel.sent).toHaveLength(0);
+    expect(tgChannel.sent[0].jid).toBe('tg:group1');
+    expect(tgChannel.sent[0].text).toBe('Here is help!');
+
+    // Step 9: Route to second channel
+    await routeOutbound([tgChannel, waChannel], 'wa:group2', 'WA response');
     expect(waChannel.sent).toHaveLength(1);
-    expect(tgChannel.sent[0].text).toBe('for telegram');
-    expect(waChannel.sent[0].text).toBe('for whatsapp');
-  });
-
-  it('findChannel returns correct channel', () => {
-    const tg = createMockChannel('tg:');
-    const wa = createMockChannel('wa:');
-
-    expect(findChannel([tg, wa], 'tg:123')).toBe(tg);
-    expect(findChannel([tg, wa], 'wa:456')).toBe(wa);
-    expect(findChannel([tg, wa], 'slack:789')).toBeUndefined();
+    expect(waChannel.sent[0].text).toBe('WA response');
   });
 });
